fix(changes): unify cross-branch diff on the typed row comparator
Extract the per-commit enumerator's typed row comparison into a shared
changes::row_compare module (OrderedRows + rows_equal) and route the
net-diff cross-branch path through it, so the codebase holds ONE definition
of "did this row change". Non-Blob columns compare by Arrow logical
equality on one-row slices (null, "", and [] stay distinct); Blob columns
compare by physical descriptor identity with a payload tie-break; only the
five reserved Lance virtual columns are skipped, so a legal _row_-prefixed
user property participates. The lossy extract_rows_with_signature and its
whole-table scan are removed; added/removed tables enumerate through the
same streamed reader. Closes the two regression cells from the prior commit.
